final code using only global state (useContext hook)
