+++ src/components/List.jsx
@@ -2,29 +2,17 @@ import { useItems } from '../context/ItemsContext';
 import AddItem from './AddItem';
 import Item from './Item';
 export default function List() {
-  const { items, setItems } = useItems();
-
-  //event handlers
-  const removeItem = (id) => {
-    setItems(items.filter((item) => item.id !== id));
-  };
-
-  const updateItem = (newItem) => {
-    setItems(items.map((item) => (item.id !== newItem.id ? item : newItem)));
-  };
-  const addItem = (item) => {
-    setItems([...items, item]);
-  };
+  const { items } = useItems();
 
   return (
     <>
-      <AddItem addItem={addItem} />
+      <AddItem />
       <hr />
       <ul>
         {items.length > 0 ? (
           items.map((item) => (
             <li key={item.id}>
-              <Item item={item} deleteItem={removeItem} updateItem={updateItem} />
+              <Item item={item} />
             </li>
           ))
         ) : (
